% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_bound.R
\name{barycenter_face}
\alias{barycenter_face}
\title{Barycenter Face}
\usage{
barycenter_face(v_list, f_list)
}
\arguments{
\item{v_list}{matrix of vertex coordinates}

\item{f_list}{matrix with 3 columns with face information.}
}
\value{
bary_mat, matrix of coordinates of barycenters of faces.
}
\description{
This function finds the barycenters of the faces of a simplicial complex given the
list of vertex coordinates and the set of faces.
}
