% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_bound.R
\name{barycenter_tet}
\alias{barycenter_tet}
\title{Barycenter Tetrahedra}
\usage{
barycenter_tet(v_list, t_list)
}
\arguments{
\item{v_list}{}

\item{t_list}{}
}
\value{
bary_mat, matrix of coordinates of barycenters of teterahedra
}
\description{
This function finds the barycenters of the tetrahedra/3-simplices of a simplicial
complex given the list of vertex coordinates and the set of tetrahedra.
}
