% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{calc_overlap_2D}
\alias{calc_overlap_2D}
\title{Calculate Overlap 2D}
\usage{
calc_overlap_2D(alpha, r = 1, rmin = 0.01, bound = "square")
}
\arguments{
\item{alpha}{radius of alpha ball}

\item{r}{length of square, radius of circle, or outer radius of annulus}

\item{rmin}{inner radius of annulus}

\item{bound}{manifold shape, options are "square", "circle", or "annulus"}
}
\value{
area of overlap
}
\description{
This function calculates the minimum coverage percentage of an alpha ball over the bounded
area being considered. 0 is no coverage, 1 means complete coverage.
For the square, r is the length of the side. For circle, r is the radius. For
the annulus, r and min_r are the two radii.
}
