% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D.R
\name{calc_overlap_3D}
\alias{calc_overlap_3D}
\title{calculate overlap in three dimensions (calc_overlap_3D)}
\usage{
calc_overlap_3D(alpha, r = 1, rmin = 0.01, bound = "cube")
}
\arguments{
\item{alpha}{radius of one sphere}

\item{r}{radius of second sphere or outer radius of shell or length of
cube side}

\item{rmin}{inner radius of shell, only needed if bound=shell}

\item{bound}{manifold type, options are "cube", "shell", and "sphere"}
}
\value{
volume of overlap
}
\description{
Calculates the volume of intersection divided by the volume of the manifold.
For the cube, r is the length of the side. For sphere, r is the radius. For
the annulus, r and min_r are the two radii.
}
