% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D.R
\name{cap_intersect_vol}
\alias{cap_intersect_vol}
\title{Intersection of spheres}
\usage{
cap_intersect_vol(alpha, r)
}
\arguments{
\item{alpha}{radius 1}

\item{r}{radius 2}
}
\value{
volume of intersection of spheres.
}
\description{
Called for sphere overlaps with alpha > r*sqrt(2). Integral precalculated and
numbers plugged in.
}
