% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{check_bound2d}
\alias{check_bound2d}
\title{Check Bounds 2D}
\usage{
check_bound2d(point, bounds)
}
\arguments{
\item{point}{1 x 2 vector or matrix holding point being checked}

\item{bounds}{bounds list from mcmc function above}
}
\value{
boolean TRUE/FALSE
}
\description{
Checks the bound parameters for a sampled point during the MCMC algorithm
}
