% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{check_bound3d}
\alias{check_bound3d}
\title{Check Bounds 3D}
\usage{
check_bound3d(point, bounds)
}
\arguments{
\item{point}{1 x 3 vector or matrix of the point being checked}

\item{bounds}{bounds list from above mcmc sampler}
}
\value{
boolean TRUE/FALSE
}
\description{
Checks the bound parameters for a sampled point during the MCMC algorithm
}
