% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_bound.R
\name{circ_face_2D}
\alias{circ_face_2D}
\title{Circumcenter face - three points in 2D
Given 3 sets of coordinates, calculates the circumcenter}
\usage{
circ_face_2D(points)
}
\arguments{
\item{points, }{3x2 matrix}
}
\value{
1x2 vector, coordinates of circumcenter
}
\description{
Circumcenter face - three points in 2D
Given 3 sets of coordinates, calculates the circumcenter
}
