% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_bound.R
\name{circ_face_3D}
\alias{circ_face_3D}
\title{Circumcenter face - three points in 3D
Given 3 sets of coordinates, calculates the circumcenter}
\usage{
circ_face_3D(points)
}
\arguments{
\item{points, }{3x3 matrix}
}
\value{
1x3 vector, coordinates of circumcenter
}
\description{
Circumcenter face - three points in 3D
Given 3 sets of coordinates, calculates the circumcenter
}
