% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_bound.R
\name{circ_tet_3D}
\alias{circ_tet_3D}
\title{Circumcenter tetrahedron - 4 points in 3D
Given 3D coordinates of 4 points, calculates circumcenter}
\usage{
circ_tet_3D(points)
}
\arguments{
\item{points, }{4x3 matrix}
}
\value{
1x3 vector, coordinates of circumcenter
}
\description{
Circumcenter tetrahedron - 4 points in 3D
Given 3D coordinates of 4 points, calculates circumcenter
}
