% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{circle_overlap_cc}
\alias{circle_overlap_cc}
\title{Circle Overlap Centered on Circumference}
\usage{
circle_overlap_cc(alpha, r = 1)
}
\arguments{
\item{alpha}{radius 1}

\item{r}{radius 2}
}
\value{
area of overlap
}
\description{
Circle overlap cc is subfunction for repeated code in calc_overlap_2D
Returns the area of two overlapping circles where one is centered on the other's
Circumference. (cc = centered on circumference )
}
