% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{circle_overlap_ia}
\alias{circle_overlap_ia}
\title{Circle Overlap Inner Annulus}
\usage{
circle_overlap_ia(alpha, R, r)
}
\arguments{
\item{alpha}{radius of circle}

\item{R}{outer radius of annulus}

\item{r}{inner radius of annulus}
}
\value{
area of overlap
}
\description{
Circle overlap ia (inner annulus) calculates area needed to subtract
when calculating area of overlap of annulus and circle.
}
