% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_bound.R
\name{circumcenter_face}
\alias{circumcenter_face}
\title{circumcenter Face}
\usage{
circumcenter_face(v_list, f_list)
}
\arguments{
\item{v_list}{matrix of vertex coordinates}

\item{f_list}{matrix with 3 columns with face information.}
}
\value{
circ_mat, matrix of coordinates of circumcenters of faces.
}
\description{
This function finds the circumcenters of the faces of a simplicial complex given the
list of vertex coordinates and the set of faces.
}
