% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_bound.R
\name{circumcenter_tet}
\alias{circumcenter_tet}
\title{circumcenter Tetrahedra}
\usage{
circumcenter_tet(v_list, t_list)
}
\arguments{
\item{v_list}{matrix of vertex coordinates}

\item{t_list}{matrix of 4 columns with tetrahedra}
}
\value{
circ_mat, matrix of coordinates of circumcenters of teterahedra
}
\description{
This function finds the circumcenters of the tetrahedra/3-simplices of a simplicial
complex given the list of vertex coordinates and the set of tetrahedra.
}
