% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_bound.R
\name{count_neighbors}
\alias{count_neighbors}
\title{Neighbors function - finds number of neighbors for each point in point cloud.}
\usage{
count_neighbors(v_list, complex)
}
\arguments{
\item{v_list}{2 or 3 column matrix}

\item{complex}{simplicial complex object}
}
\value{
n_list vector where each entry is number of neighbors for a point
}
\description{
Neighbors function - finds number of neighbors for each point in point cloud.
}
