% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{euclid_dists_point_cloud_2D}
\alias{euclid_dists_point_cloud_2D}
\title{Euclidean Distance Point Cloud 2D}
\usage{
euclid_dists_point_cloud_2D(point, point_cloud)
}
\arguments{
\item{point}{cartesian coordinates of 2D point}

\item{point_cloud}{3 column matrix with cartesian coordinates of 2D point cloud}
}
\value{
vector of distances from the point to each point in the point cloud
}
\description{
Calculates the distance matrix of a point from the point cloud.
}
