% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D.R
\name{euclid_dists_point_cloud_3D}
\alias{euclid_dists_point_cloud_3D}
\title{Euclidean Distance Point Cloud 3D}
\usage{
euclid_dists_point_cloud_3D(point, point_cloud)
}
\arguments{
\item{point}{cartesian coordinates of 3D point}

\item{point_cloud}{3 column matrix with cartesian coordinates of 3D point cloud}
}
\value{
vector of distances from the point to each point in the point cloud
}
\description{
Calculates the distance matrix of a point from the point cloud.
}
