% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex.R
\name{extract_complex_edges}
\alias{extract_complex_edges}
\title{Returns the edges of complex.}
\usage{
extract_complex_edges(complex, n_vert = 0)
}
\arguments{
\item{complex}{complex object from TDA packages}

\item{n_vert}{number of vertices in complex; default is 0, specifying
this parameter speeds up the function}
}
\value{
edge_list data frame or if empty NULL
}
\description{
Returns the edges of complex.
}
