% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex.R
\name{extract_complex_faces}
\alias{extract_complex_faces}
\title{Returns faces of complex.}
\usage{
extract_complex_faces(complex, n_vert = 0)
}
\arguments{
\item{complex}{complex object from TDA package}

\item{n_vert}{number of vertices in the complex; default is 0, specifying
this parameter speeds up function}
}
\value{
face_list data frame of points forming faces in complex
}
\description{
Returns faces of complex.
}
