% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex.R
\name{extract_complex_tet}
\alias{extract_complex_tet}
\title{Returns tetrahedra of complex (3 dimensions)}
\usage{
extract_complex_tet(complex, n_vert = 0)
}
\arguments{
\item{complex}{complex object from TDA package}

\item{n_vert}{number of vertices in the complex; default is 0, specifying this
parameter speeds up function}
}
\value{
tet_list data frame of points forming tetrahedra in complex
}
\description{
Returns tetrahedra of complex (3 dimensions)
}
