% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_bound.R
\name{extreme_pts}
\alias{extreme_pts}
\title{Extreme points
Finds the boundary points of a simplicial complex}
\usage{
extreme_pts(complex, n_vert, dimension)
}
\arguments{
\item{complex}{complex list object}

\item{n_vert}{number of vertices in the complex}

\item{dimension}{number, 2 or 3}
}
\value{
vector of all vertices on the boundary
}
\description{
Extreme points
Finds the boundary points of a simplicial complex
}
