% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{generate_ashape2d}
\alias{generate_ashape2d}
\title{Generate 2D alpha shape}
\usage{
generate_ashape2d(
  point_cloud,
  N,
  tau,
  delta = 0.01,
  bound = "circle",
  mu = NULL,
  sig = NULL
)
}
\arguments{
\item{point_cloud}{2 column matrix of all points from all shapes in initial
data set}

\item{N}{number of shapes in initial data set}

\item{tau}{tau bound}

\item{delta}{probability of not preserving homology; default is 0.01}

\item{bound}{manifold bound being used to sample points}

\item{mu}{mean of truncated distribution from which alpha sampled; default tau/3}

\item{sig}{standard deviation of truncated distribution from which alpha
sampled; default tau/12}
}
\value{
new_ashape two dimensional alpha shape object from alphahull library
}
\description{
Generate 2D alpha shape
}
