% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc.R
\name{generate_ashape3d}
\alias{generate_ashape3d}
\title{Generate 3D alpha shape}
\usage{
generate_ashape3d(
  point_cloud,
  N,
  tau,
  delta = 0.05,
  bound = "sphere",
  mu = NULL,
  sig = NULL
)
}
\arguments{
\item{point_cloud}{3 column matrix of all points from all shapes in initial
data set}

\item{N}{number of shapes in initial data set}

\item{tau}{tau bound}

\item{delta}{probability of not preserving homology; default is 0.05}

\item{bound}{manifold bound being used to sample points}

\item{mu}{mean of truncated distribution from which alpha sampled; default tau/3}

\item{sig}{standard deviation of truncated distribution from which alpha
sampled; default tau/12}
}
\value{
new_ashape three dimensional alpha shape object from alphashape3d library
}
\description{
Generate 3D alpha shape
}
