% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complex.R
\name{get_alpha_complex}
\alias{get_alpha_complex}
\title{Get alpha complex}
\usage{
get_alpha_complex(points, alpha)
}
\arguments{
\item{points}{point cloud for alpha complex, in form of 2 column of 3 column
matrix with nonzero number of rows}

\item{alpha}{alpha parameter for building the alpha complex}
}
\value{
complex list of vertices, edges, faces, and tetrahedra.
}
\description{
Generates alpha complex for a set of points and parameter alpha
}
