% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{get_area}
\alias{get_area}
\title{Get area}
\usage{
get_area(r, rmin, bound)
}
\arguments{
\item{r}{side length (square) or radius (circle, annulus)}

\item{rmin}{radius of inner circle for annulus}

\item{bound}{square, circle, or annulus}
}
\value{
area, number
}
\description{
Quickly calculate which area needed for a homology bound; here to clean up
code above
}
