% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{get_volume}
\alias{get_volume}
\title{Get volume}
\usage{
get_volume(r, rmin, bound)
}
\arguments{
\item{r}{side length (cube) or radius (sphere, shell)}

\item{rmin}{radius of inner sphere for shell}

\item{bound}{cube, sphere, shell}
}
\value{
volume, number
}
\description{
Quickly calculate which volume needed for a homology bound; here to clean up
code above
}
