% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_list.R
\docType{data}
\name{m_list}
\alias{m_list}
\title{Microcebus Teeth as objects for ashapesampler}
\format{
A named list containing Microcebus molars as mesh3d objects of the Rvcg package
\describe{
\item{Vertices}{3 x n matrix containing n vertices as Euclidean coordinates}
\item{cmplx}{list of vertices, edges, and faces forming simplicial complex}
}
}
\source{
vignette/teeth_demo_3D.Rmd
}
\usage{
m_list
}
\description{
This data set is of 8 of Microcebus mandibular molars characterized as vertices
and the resulting simplicial complex, as used in
Winn-Nunez, et. al. (2023). These were imported from off files with
function readOFF from the ashapesampler package.
}
\keyword{datasets}
