% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{n_bound_connect_2D}
\alias{n_bound_connect_2D}
\title{n Bound Connect 2D}
\usage{
n_bound_connect_2D(alpha, delta = 0.05, r = 1, rmin = 0.01, bound = "square")
}
\arguments{
\item{alpha}{alpha parameter for alpha shape}

\item{delta}{probability of isolated point}

\item{r}{length of square, radius of circle, or outer radius of annulus}

\item{rmin}{inner radius of annulus}

\item{bound}{manifold shape, options are "square", "circle", or "annulus"}
}
\value{
minimum number of points to meet probability threshold.
}
\description{
This is the bound for connectivity based on samples.
}
