% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{n_bound_homology_2D}
\alias{n_bound_homology_2D}
\title{n Bound Homology 2D}
\usage{
n_bound_homology_2D(area, epsilon, tau = 1, delta = 0.05)
}
\arguments{
\item{area}{area of manifold from which points being sampled}

\item{epsilon}{size of balls of cover}

\item{tau}{number bound}

\item{delta}{probability of not recovering homology}
}
\value{
n, number of points needed
}
\description{
#' Function returns the minimum number of points to preserve the homology with
an open cover of radius alpha.
}
