% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D.R
\name{n_bound_homology_3D}
\alias{n_bound_homology_3D}
\title{n Bound Homology 3D}
\usage{
n_bound_homology_3D(volume, epsilon, tau = 1, delta = 0.05)
}
\arguments{
\item{volume}{volume of manifold from which points being sampled}

\item{epsilon}{size of balls of cover}

\item{tau}{number bound}

\item{delta}{probability of not recovering homology}
}
\value{
n, number of points needed
}
\description{
Calculates number of points needed to be samped from manifold for open
ball cover to have same homology as original manifold. See Niyogi et al 2008
}
