% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2D.R
\name{plot_2D_ashape}
\alias{plot_2D_ashape}
\title{Plot 2D alpha shape in ggplot2}
\usage{
plot_2D_ashape(my_ashape, color1 = "gray", color2 = "blue")
}
\arguments{
\item{my_ashape}{two dimensional alpha shape}

\item{color1}{(optional) color of interior, default gray}

\item{color2}{(optional) color of boundary edges/points, default blue}
}
\value{
nothing; plots shape.
}
\description{
Plot 2D alpha shape in ggplot2
}
