% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_read_write.R
\name{read_alpha_txt}
\alias{read_alpha_txt}
\title{Read alpha text file}
\usage{
read_alpha_txt(file_name)
}
\arguments{
\item{file_name}{name and path of file to be read. File is of format output
by write_alpha_txt function}
}
\value{
alpha shape object
}
\description{
Read alpha text file
}
