% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_read_write.R
\name{readOFF}
\alias{readOFF}
\title{Read OFF File}
\usage{
readOFF(file_name)
}
\arguments{
\item{file_name}{path and name of file to be read}
}
\value{
complex_info list object containing two components, "Vertices" which
holds the vertex coordinates and "cmplx" which holds the complex list
object.
}
\description{
This is a function to read OFF files for triangular meshes into the form
that is required to use other functions in the package.
}
