% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D.R
\name{runif_ball_3D}
\alias{runif_ball_3D}
\title{Uniform Ball 3D}
\usage{
runif_ball_3D(n, r = 1)
}
\arguments{
\item{n}{number of points}

\item{r}{radius of ball, default r=1}
}
\value{
n by 3 matrix of points
}
\description{
Returns points uniformly centered from closed ball of radius r in 3D space
}
\examples{
# Sample 100 points from unit ball
runif_ball_3D(100)
# Sample 100 points from ball of radius 0.5
runif_ball_3D(100, r=0.5)
}
