% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D.R
\name{runif_cube}
\alias{runif_cube}
\title{r Uniform Cube}
\usage{
runif_cube(n, xmin = 0, xmax = 1, ymin = 0, ymax = 1, zmin = 0, zmax = 1)
}
\arguments{
\item{n}{number of points to be sampled}

\item{xmin}{miniumum x coordinate}

\item{xmax}{maximum x coordinate}

\item{ymin}{minimum y coordinate}

\item{ymax}{maximum y coordinate}

\item{zmin}{minimum z coordinate}

\item{zmax}{maximum z coordinate}
}
\value{
n by 3 matrix of points
}
\description{
Returns points uniformly sampled from cube or rectangular prism in space.
}
\examples{
# Sample 100 points from unit cube
runif_cube(100)
# Sample 100 points from unit cube centered on origin
runif_cube(100, 0.5, 0.5, 0.5, 0.5, 0.5, 0.5)
}
