% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{runif_disk}
\alias{runif_disk}
\title{Uniform sampling from disk}
\usage{
runif_disk(n, r = 1)
}
\arguments{
\item{n}{number of points to sample}

\item{r}{radius of disk}
}
\value{
points n by 2 matrix of points sampled
}
\description{
Returns points uniformly sampled from disk of radius r in plane
}
\examples{
# Sample 100 points from unit disk
runif_disk(100)
# Sample 100 points from disk of radius 0.7
runif_disk(100, 0.7)
}
