% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D.R
\name{runif_shell_3D}
\alias{runif_shell_3D}
\title{Uniform Shell 3D}
\usage{
runif_shell_3D(n, rmax = 1, rmin = 0.5)
}
\arguments{
\item{n}{number of points}

\item{rmax}{radius of outer sphere}

\item{rmin}{radius of inner sphere}
}
\value{
n by 3 matrix of points
}
\description{
Returns points uniformly sampled from spherical shell in 3D
}
\examples{
# Sample 100 points with defaults rmax=1, rmin=0.5
runif_shell_3D(100)
# Sample 100 points with rmax=0.75, rmin=0.25
runif_shell_3D(100, 0.75, 0.25)
}
