% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{runif_square}
\alias{runif_square}
\title{Uniform Sampling from Square}
\usage{
runif_square(n, xmin = 0, xmax = 1, ymin = 0, ymax = 1)
}
\arguments{
\item{n}{number of points}

\item{xmin}{minimum x coordinate}

\item{xmax}{maximum x coordinate}

\item{ymin}{minimum y coordinate}

\item{ymax}{maximum y coordinate}
}
\value{
n by 2 matrix of points
}
\description{
Returns points uniformly sampled from square or rectangle in plane.
}
\examples{
# Sample 100 points from unit square
runif_square(100)
# Sample 100 points from unit square centered at origin
runif_square(100, 0.5, 0.5, 0.5, 0.5)
}
