% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D.R
\name{sphere_overlap_cs}
\alias{sphere_overlap_cs}
\title{sphere overlap when one is centered on circumference of the other}
\usage{
sphere_overlap_cs(alpha, r)
}
\arguments{
\item{alpha}{radius 1}

\item{r}{radius 2}
}
\value{
volume of intersection
}
\description{
Sphere overlap cs is subfunction for repeated code in calc_overlap_3D
Returns the area of two overlapping spheres where one is centered on the
other's surface (cs = centered on surface)
}
