% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D.R
\name{sphere_overlap_is}
\alias{sphere_overlap_is}
\title{sphere overlap inner shell}
\usage{
sphere_overlap_is(alpha, rmax, rmin)
}
\arguments{
\item{alpha}{radius of sphere}

\item{rmax}{outer radius of shell}

\item{rmin}{inner radius of shell}
}
\value{
volume of intersection
}
\description{
Sphere overlap is (inner shell) calculates area needed to subtract
when calculating volume of overlap of shell and sphere.
}
