% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D.R
\name{spherical_cap}
\alias{spherical_cap}
\title{Spherical cap}
\usage{
spherical_cap(r, h)
}
\arguments{
\item{r}{radius}

\item{h}{height of cap}
}
\value{
v_c volume of spherical cap
}
\description{
Calculates the volume of a sphere cap given radius r and height of cap h
}
