% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_mesh.R
\docType{data}
\name{t_mesh}
\alias{t_mesh}
\title{Tarsius Teeth as mesh3d objects}
\format{
A named list containing Microcebus molars as mesh3d objects of the Rvcg package
\describe{
\item{vb}{4 x n matrix containing n vertices as homolougous coordinates}
\item{it}{3 x m matrix containing vertex indices forming triangular faces}
\item{normals}{4 x n matrix containing vertex normals (homologous coordinates)}
}
}
\source{
\url{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/K9A0EG&faces-redirect=true}
}
\usage{
t_mesh
}
\description{
This data set is of 7 mesh3d representations of Tarsius mandibular molars
use in Winn-Nunez, et. al. (2023). These were imported from off files using
the vcgImport function from package Rvcg.
See \url{https://CRAN.R-project.org/package=Rvcg}
for more information about the mesh3d object and the vcgImport function.
}
\keyword{datasets}
