% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_bound.R
\name{tau_bound}
\alias{tau_bound}
\title{tau_bound}
\usage{
tau_bound(v_list, e_list, extremes = NULL, f_list = NULL, t_list = NULL)
}
\arguments{
\item{v_list}{matrix or data frame of cartesian coordinates of vertices in
in point cloud}

\item{e_list}{data frame or matrix of edges, listed as pairs of vertices}

\item{extremes}{matrix or data frame of cartesian coordinates of vertices on
the boundary of the data frame. If no list given, function will
assume all points are extreme and check them all. Inclusion of
this parameter speeds up the process.}

\item{f_list}{data frame or matrix of faces, listed as sets of 3 vertices}

\item{t_list}{data frame or matrix of faces, listed as sets of 4 vertices}
}
\value{
tau, real nonnegative number.
}
\description{
This function finds the bound of tau, which is the maximum length of the
fiber bundle off of a shape for determining the density of points necessary
to recover the homology from the open cover. See Niyogi et al 2008. Function
checks length of edges and distances to barycenters from each vertex before
checking against the rest of the point cloud and finds the minimum length.
}
