% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_read_write.R
\name{write_alpha_txt}
\alias{write_alpha_txt}
\title{Write Alpha Text file}
\usage{
write_alpha_txt(ashape, file_name)
}
\arguments{
\item{ashape}{alpha shape object, can be 2D or 3D alpha shape}

\item{file_name}{path and name of file to create and write text to}
}
\value{
does not return anything; writes file that can be read back to R via
read_alpha_txt
}
\description{
Write Alpha Text file
}
